[code-infra] Move `isDeepEqual` to @mui/x-internals (#17129)
