+++ packages/x-data-grid-generator/src/hooks/useQuery.ts
@@ -1,7 +1,7 @@
 'use client';
 import * as React from 'react';
 import { getGridDefaultColumnTypes, GridRowModel } from '@mui/x-data-grid-pro';
-import { isDeepEqual } from '@mui/x-data-grid/internals';
+import { isDeepEqual } from '@mui/x-internals/isDeepEqual';
 import {
   useDemoData,
   UseDemoDataOptions,
