+++ packages/x-data-grid/src/utils/utils.ts
@@ -42,142 +42,6 @@ export function range(from: number, to: number) {
   return Array.from({ length: to - from }).map((_, i) => from + i);
 }
 
-/**
- * Based on `fast-deep-equal`
- *
- * MIT License
- *
- * Copyright (c) 2017 Evgeny Poberezkin
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- * We only type the public interface to avoid dozens of `as` in the function.
- */
-export function isDeepEqual<T>(actual: any, expected: T): actual is T;
-
-export function isDeepEqual(a: any, b: any) {
-  if (a === b) {
-    return true;
-  }
-
-  if (a && b && typeof a === 'object' && typeof b === 'object') {
-    if (a.constructor !== b.constructor) {
-      return false;
-    }
-
-    if (Array.isArray(a)) {
-      const length = a.length;
-      if (length !== b.length) {
-        return false;
-      }
-      for (let i = 0; i < length; i += 1) {
-        if (!isDeepEqual(a[i], b[i])) {
-          return false;
-        }
-      }
-      return true;
-    }
-
-    if (a instanceof Map && b instanceof Map) {
-      if (a.size !== b.size) {
-        return false;
-      }
-
-      const entriesA = Array.from(a.entries());
-
-      for (let i = 0; i < entriesA.length; i += 1) {
-        if (!b.has(entriesA[i][0])) {
-          return false;
-        }
-      }
-      for (let i = 0; i < entriesA.length; i += 1) {
-        const entryA = entriesA[i];
-        if (!isDeepEqual(entryA[1], b.get(entryA[0]))) {
-          return false;
-        }
-      }
-      return true;
-    }
-
-    if (a instanceof Set && b instanceof Set) {
-      if (a.size !== b.size) {
-        return false;
-      }
-
-      const entries = Array.from(a.entries());
-      for (let i = 0; i < entries.length; i += 1) {
-        if (!b.has(entries[i][0])) {
-          return false;
-        }
-      }
-      return true;
-    }
-
-    if (ArrayBuffer.isView(a) && ArrayBuffer.isView(b)) {
-      const length = (a as any).length;
-      if (length !== (b as any).length) {
-        return false;
-      }
-      for (let i = 0; i < length; i += 1) {
-        if ((a as any)[i] !== (b as any)[i]) {
-          return false;
-        }
-      }
-      return true;
-    }
-
-    if (a.constructor === RegExp) {
-      return a.source === b.source && a.flags === b.flags;
-    }
-    if (a.valueOf !== Object.prototype.valueOf) {
-      return a.valueOf() === b.valueOf();
-    }
-    if (a.toString !== Object.prototype.toString) {
-      return a.toString() === b.toString();
-    }
-
-    const keys = Object.keys(a);
-    const length = keys.length;
-    if (length !== Object.keys(b).length) {
-      return false;
-    }
-
-    for (let i = 0; i < length; i += 1) {
-      if (!Object.prototype.hasOwnProperty.call(b, keys[i])) {
-        return false;
-      }
-    }
-
-    for (let i = 0; i < length; i += 1) {
-      const key = keys[i];
-      if (!isDeepEqual(a[key], b[key])) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  // true if both NaN, false otherwise
-  // eslint-disable-next-line no-self-compare
-  return a !== a && b !== b;
-}
-
 // Pseudo random number. See https://stackoverflow.com/a/47593316
 function mulberry32(a: number): () => number {
   return () => {
