+++ packages/x-internals/src/isDeepEqual/index.ts
@@ -1,0 +1,1 @@
+export * from './isDeepEqual';
